+++ tests/test_skoolhtml.py
@@ -1351,6 +1351,17 @@ class SkoolMacroTest(HtmlWriterTestCase, CommonSkoolMacroTest):
         output = writer.expand('#LINK:CustomPage#40000(foo)', ASMDIR)
         self._assert_link_equals(output, '../CustomPage.html#40000', 'foo')
 
+    def test_macro_link_to_box_page_item_with_blank_link_text(self):
+        ref = '\n'.join((
+            '[Page:Boxes]',
+            'SectionPrefix=Box',
+            '[Box:item1:Item 1]',
+            'This is item 1.'
+        ))
+        writer = self._get_writer(ref=ref)
+        output = writer.expand('#LINK:Boxes#item1()', ASMDIR)
+        self._assert_link_equals(output, '../Boxes.html#item1', 'Item 1')
+
     def test_macro_link_invalid(self):
         writer, prefix = CommonSkoolMacroTest.test_macro_link_invalid(self)
         self._assert_error(writer, '#LINK:nonexistentPageID(text)', 'Unknown page ID: nonexistentPageID', prefix)